- 테스트 케이스 이전 - 테스트 파일 변경 - 모듈화
