+++ index.js
@@ -1,193 +1,126 @@
-var assert = require('assert');
-var chai = require('chai');
-var expect = chai.expect;
 class Province {
-    constructor(doc) {
-        this._name = doc.name
-        this._producers = [];
-        this._totalProduction = 0;
-        this._demand = doc.demand;
-        this._price = doc.price;
-        doc.producers.forEach( d =>  this.addProducer(new Producer(this, d)));
-            
-    }
+  constructor(doc) {
+    this._name = doc.name
+    this._producers = [];
+    this._totalProduction = 0;
+    this._demand = doc.demand;
+    this._price = doc.price;
+    doc.producers.forEach(d => this.addProducer(new Producer(this, d)));
+  }
 
-    addProducer(arg) {
-        this._producers.push(arg);
-        this._totalProduction += arg.production;
-    }
+  addProducer(arg) {
+    this._producers.push(arg);
+    this._totalProduction += arg.production;
+  }
 
-    get name() {
-        return this._name;
-    }
-    
-    get producers() {
-        return this._producers.slice();
-    }
+  get name() {
+    return this._name;
+  }
 
-    get totalProduction() {
-        return this._totalProduction;
-    }
+  get producers() {
+    return this._producers.slice();
+  }
 
-    set totalProduction(arg) {
-        this._totalProduction = arg;
-    }
+  get totalProduction() {
+    return this._totalProduction;
+  }
 
-    get demand() {
-        return this._demand;
-    }
+  set totalProduction(arg) {
+    this._totalProduction = arg;
+  }
 
-    set demand(arg) {
-        this._demand = parseInt(arg);
-    }
+  get demand() {
+    return this._demand;
+  }
 
-    get price() {
-        return this._price;
-    }
+  set demand(arg) {
+    this._demand = parseInt(arg);
+  }
 
-    set price(arg) {
-        this._price = parseInt(arg);
-    }
+  get price() {
+    return this._price;
+  }
 
-    get shortFall() {
-        return this._demand - this.totalProduction;
-    }
+  set price(arg) {
+    this._price = parseInt(arg);
+  }
 
-    get profit() {
-        return this.demandValue - this.demandCost;
-    }
+  get shortFall() {
+    return this._demand - this.totalProduction;
+  }
 
-    get demandValue() {
-        return this.satisfiedDemand * this.price;
-    }
+  get profit() {
+    return this.demandValue - this.demandCost;
+  }
 
-    get satisfiedDemand() {
-        return Math.min(this._demand, this.totalProduction);
-    }
+  get demandValue() {
+    return this.satisfiedDemand * this.price;
+  }
 
-    get demandCost() {
-        let remainingDemand = this.demand;
-        let result = 0;
-        this.producers
-        .sort((a,b) => a.cost - b.cost)
-        .forEach( p => {
-            const contribution = Math.min(remainingDemand, p.production);
-            remainingDemand -= contribution;
-            result += contribution * p.cost;
-        });
-        return result;
-    }
+  get satisfiedDemand() {
+    return Math.min(this._demand, this.totalProduction);
+  }
+
+  get demandCost() {
+    let remainingDemand = this.demand;
+    let result = 0;
+    this.producers
+      .sort((a, b) => a.cost - b.cost)
+      .forEach(p => {
+        const contribution = Math.min(remainingDemand, p.production);
+        remainingDemand -= contribution;
+        result += contribution * p.cost;
+      });
+    return result;
+  }
 }
 
 function sampleProvinceData() {
-    return {
-        name: "Asia",
-        producers: [
-            {name: "Byzantium", cost: 10, production: 9},
-            {name: "Attalia", cost: 12, production: 10},
-            {name: "Sinope", cost: 10, production: 6},
-        ],
-        demand: 30,
-        price: 20
-    };
+  return {
+    name: "Asia",
+    producers: [
+      {name: "Byzantium", cost: 10, production: 9},
+      {name: "Attalia", cost: 12, production: 10},
+      {name: "Sinope", cost: 10, production: 6},
+    ],
+    demand: 30,
+    price: 20
+  };
 }
 
 class Producer {
-    constructor(aProvince, data) {
-        this._province = aProvince;
-        this._cost = data.cost;
-        this._name = data.name;
-        this._production = data.production || 0;
-    }
+  constructor(aProvince, data) {
+    this._province = aProvince;
+    this._cost = data.cost;
+    this._name = data.name;
+    this._production = data.production || 0;
+  }
 
-    get name() {
-        return this._name;
-    }
+  get name() {
+    return this._name;
+  }
 
-    get cost() {
-        return this._cost;
-    }
+  get cost() {
+    return this._cost;
+  }
 
-    set cost(arg) {
-        this._cost = parseInt(arg);
-    }
+  set cost(arg) {
+    this._cost = parseInt(arg);
+  }
 
-    get production() {
-        return this._production;
-    }
+  get production() {
+    return this._production;
+  }
 
-    set production(amountStr) {
-        const amount = parseInt(amountStr);
-        const newProduction = Number.isNaN(amount) ? 0: amount;
-        this._province.totalProduction += newProduction - this._production;
-        this._production = newProduction;
-    }
+  set production(amountStr) {
+    const amount = parseInt(amountStr);
+    const newProduction = Number.isNaN(amount) ? 0 : amount;
+    this._province.totalProduction += newProduction - this._production;
+    this._production = newProduction;
+  }
 }
 
-
-// test case 
-describe('province', function() {
-    let asia;
-    beforeEach(function () {
-        asia = new Province(sampleProvinceData());
-    })
-    it('shortfall', function() {
-        expect(asia.shortFall).equal(5);
-    });
-    it('profit', function() {
-        expect(asia.profit).equal(230);
-    })
-    it('change production', function() {
-        asia.producers[0].production = 20;
-        expect(asia.shortFall).equal(-6);
-        expect(asia.profit).equal(292);
-    })
-    it('zero demand', function() {
-        asia.demand = 0;
-        expect(asia.shortFall).equal(-25);
-        expect(asia.profit).equal(0);
-
-    })
-    it('negative demand', function() {
-        asia.demand = -1;
-        expect(asia.shortFall).equal(-26)
-        expect(asia.profit).equal(-10)
-    })
-    it('empty string demand', function() {
-        asia.demand = "";
-        expect(asia.shortFall).NaN
-        expect(asia.profit).NaN
-    })
-});
-
-describe('no producers', function() {
-    let noProducers;
-    beforeEach(function() {
-        const data = {
-            name: "No producers",
-            producers: [],
-            demand: 30,
-            price: 20
-        };
-        noProducers = new Province(data);
-    });
-    it('shortfall', function() {
-        expect(noProducers.shortFall).equal(30);
-    });
-    it('profit', function() {
-        expect(noProducers.profit).equal(0);
-    });
-})
-
-describe('string for producers', function() {
-    it('', function() {
-        const data = {
-            name: "String producers",
-            producers: "",
-            demand: 30,
-            price: 20
-        };
-        const prov = new Province(data);
-        expect(prov.shortFall).equal(0);
-    });
-})
+module.exports = {
+  Province,
+  sampleProvinceData
+}
