+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "test": "mocha index.js -w"
+    "test": "mocha test.js -w"
   },
   "dependencies": {
     "chai": "^4.3.4"
